pyproject.toml
[tool.poetry]
name = "transpire"
version = "0.1.0"
description = "the OCF Kubernetes helper library"
authors = ["Nikhil Jha <njha@ocf.berkeley.edu>"]
license = "LGPL"

[tool.poetry.scripts]
transpire = 'transpire.__main__:cli'

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0" # make sure to bump types-PyYAML as well
kubernetes = "^23.3.0" # make sure to bump kubernetes-typed as well
tomlkit = "^0.10.0"
click = "^8.0.4"
loguru = "^0.6.0"
better-exceptions = "^0.3.3"
inquirer = "^2.9.1"
pydantic = "^1.9.0"
hera-workflows = "^4.0.1"
hvac = "^1.0.2"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}
mypy = "^0.940"
types-PyYAML = "^6.0.4"
kubernetes-typed = { git = "https://github.com/nikhiljha/kubernetes-typed.git", tag = "v23.3.0" }
types-requests = "^2.27.19"
flake8 = "^4.0.1"
isort = "^5.10.1"
pre-commit = "^2.18.1"
pytest = "^7.1.2"
coverage = { version = "^6.3.2", extras = ["toml"] }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.coverage.run]
source = ["transpire"]
branch = true
